Refactor  validation in RO
